Move code classes to dedicated directory
