+++ src/classes/es-constructor.spec.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it } from '@jest/globals';
 import { jsStringLiteral } from 'httongue';
-import { EsCode } from '../es-code.js';
-import { esline } from '../esline.tag.js';
+import { EsCode } from '../code/es-code.js';
+import { esline } from '../code/esline.tag.js';
 import { EsSignature } from '../functions/es-signature.js';
 import { EsBundle } from '../scopes/es-bundle.js';
 import { EsClass } from './es-class.js';
