+++ src/classes/es-field.ts
@@ -1,5 +1,5 @@
-import { EsSnippet } from '../es-snippet.js';
-import { esline } from '../esline.tag.js';
+import { EsSnippet } from '../code/es-snippet.js';
+import { esline } from '../code/esline.tag.js';
 import { EsClass } from './es-class.js';
 import { EsMember, EsMemberRef } from './es-member.js';
