+++ src/functions/es-function.ts
@@ -1,5 +1,5 @@
-import { EsSnippet } from '../es-snippet.js';
-import { esline } from '../esline.tag.js';
+import { EsSnippet } from '../code/es-snippet.js';
+import { esline } from '../code/esline.tag.js';
 import { esFunctionOrBundle } from '../impl/es-function-or-bundle.js';
 import {
   EsDeclarationContext,
