+++ src/mod.ts
@@ -2,12 +2,8 @@
  * @module esgen
  */
 export * from './classes/mod.js';
+export * from './code/mod.js';
 export * from './declarations/mod.js';
-export * from './es-code.js';
-export * from './es-comment.js';
-export * from './es-output.js';
-export * from './es-snippet.js';
-export * from './esline.tag.js';
 export * from './functions/mod.js';
 export * from './scopes/mod.js';
 export * from './symbols/mod.js';
